Refactor module layout
