+++ tests/test_sample.rs
@@ -1,8 +1,10 @@
+use deptreevis::walker;
+
 #[cfg(test)]
 mod tests {
     #[test]
-    fn sample(){
-	let thing = 2;
-	assert_eq!(thing, 2);
+    fn sample() {
+        let thing = 2;
+        assert_eq!(thing, 2);
     }
 }
